% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_vortex.R
\name{generate_vortex}
\alias{generate_vortex}
\title{Generate Vortex}
\usage{
generate_vortex(n = 25, s = 1234)
}
\arguments{
\item{n}{Number of points. Default 25.}

\item{s}{Seed value. Default 1234.}
}
\value{
A tibble.
}
\description{
This function generates data for the vortex art series
}
