% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_waves.R
\name{generate_waves}
\alias{generate_waves}
\title{Generate Waves}
\usage{
generate_waves(a = 23, b = 6, s = 2021)
}
\arguments{
\item{a}{sine wave parameter. Default 23.}

\item{b}{Cosine wave parameter. Default 6.}

\item{s}{Seed value. Default 2021.}
}
\value{
A ggplot object.
}
\description{
This function generates a generative art ggplot object
from sine and cosine waves.
}
