% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vortex}
\alias{vortex}
\title{Vortex data set.}
\format{
A tibble with 250 rows and 3 variables:
\describe{
\item{id}{id value}
\item{value}{random integer}
\item{type}{type for colour}
}
}
\source{
\url{https://github.com/nrennie/aRtData/tree/main/R/generate_vortex.R}
}
\usage{
vortex
}
\description{
A data set containing random data for the vortex art series
}
\keyword{datasets}
