% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waves}
\alias{waves}
\title{Waves data set.}
\format{
A tibble with 15708 rows and 2 variables:
\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
}
}
\source{
\url{https://github.com/nrennie/aRtData/tree/main/R/generate_waves.R}
}
\usage{
waves
}
\description{
A data set containing random data for the waves art series
}
\keyword{datasets}
